fake_header_methods moved to WebTestBase and ResourceTestBase
