+++ servicex_app/tests/resource_test_base.py
@@ -29,6 +29,7 @@ from datetime import datetime
 from unittest.mock import MagicMock
 
 from flask.testing import FlaskClient
+from flask_jwt_extended import create_access_token
 from pytest import fixture
 from servicex import create_app
 from servicex.code_gen_adapter import CodeGenAdapter
@@ -39,6 +40,14 @@ from servicex.rabbit_adaptor import RabbitAdaptor
 
 
 class ResourceTestBase:
+
+    @staticmethod
+    def fake_header():
+        access_token = create_access_token('testuser')
+        headers = {
+            'Authorization': 'Bearer {}'.format(access_token)
+        }
+        return headers
 
     @staticmethod
     def _app_config():
