+++ servicex_app/tests/resources/transformation/test_get_all.py
@@ -26,7 +26,6 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 from flask import Response
-from flask_jwt_extended import create_access_token
 from pytest import fixture
 from tests.resource_test_base import ResourceTestBase
 
@@ -35,14 +34,6 @@ class TestAllTransformationRequest(ResourceTestBase):
     @staticmethod
     def example_json():
         return [{'request_id': '123'}, {'request_id': '456'}]
-
-    @staticmethod
-    def fake_header():
-        access_token = create_access_token('testuser')
-        headers = {
-            'Authorization': 'Bearer {}'.format(access_token)
-        }
-        return headers
 
     @fixture()
     def mock_return_json(self, mocker):
