+++ servicex_app/tests/web/web_test_base.py
@@ -30,10 +30,19 @@ from datetime import datetime
 from flask import template_rendered
 from flask.testing import FlaskClient
 from pytest import fixture
+from flask_jwt_extended import create_access_token
 
 
 class WebTestBase:
     module = ""
+
+    @staticmethod
+    def fake_header():
+        access_token = create_access_token('testuser')
+        headers = {
+            'Authorization': 'Bearer {}'.format(access_token)
+        }
+        return headers
 
     @staticmethod
     def _app_config():
@@ -129,23 +138,23 @@ class WebTestBase:
             "last_authentication": 1595620302,
             "identity_set": [
                 {
-                  "email": "jane@cern.ch",
-                  "identity_provider_display_name": "CERN",
-                  "identity_provider": "primary-identity-provider-id",
-                  "organization": "CERN",
-                  "username": "jane@cern.ch",
-                  "name": "Jane Doe",
-                  "last_authentication": 1595620302,
-                  "sub": "primary-oauth-id"
+                    "email": "jane@cern.ch",
+                    "identity_provider_display_name": "CERN",
+                    "identity_provider": "primary-identity-provider-id",
+                    "organization": "CERN",
+                    "username": "jane@cern.ch",
+                    "name": "Jane Doe",
+                    "last_authentication": 1595620302,
+                    "sub": "primary-oauth-id"
                 },
                 {
-                  "email": "jane@uchicago.edu",
-                  "identity_provider_display_name": "Google",
-                  "last_authentication": 1595552908,
-                  "identity_provider": "secondary-oauth-id",
-                  "username": "jane@uchicago.edu@accounts.google.com",
-                  "name": "Jane Doe",
-                  "sub": "secondary-oauth-id"
+                    "email": "jane@uchicago.edu",
+                    "identity_provider_display_name": "Google",
+                    "last_authentication": 1595552908,
+                    "identity_provider": "secondary-oauth-id",
+                    "username": "jane@uchicago.edu@accounts.google.com",
+                    "name": "Jane Doe",
+                    "sub": "secondary-oauth-id"
                 }
             ],
             "name": "Jane Doe",
